+++ Editor/Attributes.cs
@@ -3,15 +3,6 @@ using UnityEngine;
 
 namespace UnityEditor.SettingsManagement
 {
-#if !UNITY_2018_3_OR_NEWER
-    public enum SettingsScope
-    {
-        Any = 0x0,
-        User = 0x1,
-        Project = 0x2
-    }
-#endif
-
     /// <summary>
     /// Register a static field of type IUserSetting with the UserSettingsProvider window.
     /// </summary>