+++ Samples~/PackageWithProjectAndUserSettings/MySettingsProvider.cs
@@ -3,9 +3,15 @@ using UnityEngine;
 
 namespace UnityEditor.SettingsManagement.Examples
 {
+	/// <summary>
+	/// To create an entry in the Preferences window, define a new SettingsProvider inheriting <see cref="UserSettingsProvider"/>.
+	/// You can also choose to implement your own SettingsProvider and ignore this implementation. The benefit of using
+	/// <see cref="UserSettingsProvider"/> is that all <see cref="UserSetting{T}"/> fields in the assembly are automatically
+	/// populated within the preferences, with support for search and resetting default values.
+	/// </summary>
 	static class MySettingsProvider
 	{
-		const string k_PreferencesPath = "Preferences/My Settings";
+		const string k_PreferencesPath = "Preferences/Package With Project and User Settings";
 
 #if UNITY_2018_3_OR_NEWER
 		[SettingsProvider]
@@ -19,6 +25,7 @@ namespace UnityEditor.SettingsManagement.Examples
 		}
 #else
 
+	// For backwards compatibility it is possible to create an instance of UserSettingsProvider and invoke OnGUI manually.
 	[NonSerialized]
 	static UserSettingsProvider s_SettingsProvider;