+++ Samples~/PackageWithProjectAndUserSettings/MySettingsWrapper.cs
@@ -2,6 +2,8 @@
 
 namespace UnityEditor.SettingsManagement.Examples
 {
+    // Usually you will only have a single Settings instance, so it is convenient to define a UserSetting<T> implementation
+    // that points to your instance. In this way you avoid having to pass the Settings parameter in setting field definitions.
     class MySetting<T> : UserSetting<T>
     {
         public MySetting(string key, T value, SettingsScope scope = SettingsScope.Project)