+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.0-preview.8] - 2019-03-29
+
+### Features
+
+- Support saving multiple settings repositories within a project
+
+### Changes
+
+- Rename `ProjectSettingsRepository` -> `PackageSettingsRepository`.
+- Update readme with a complete code example.
+- Add additional documentation and unit tests.
+- Setting repositories now have names.
+
+### Bug Fixes
+
+- Fixed missing gear icon in Settings Provider implementation.
+
 ## [0.1.0-preview.4] - 2019-02-28
 
 ### Package configuration update.